feat: integrate exceptions into services, controllers and repositories

+++ src/main/java/com/uade/tpo/ecommerce/service/CartItemService.java
@@ -1,6 +1,8 @@
 package com.uade.tpo.ecommerce.service;
 
 import com.uade.tpo.ecommerce.entity.CartItem;
+import com.uade.tpo.ecommerce.exceptions.CartItemNotFoundException;
+import com.uade.tpo.ecommerce.exceptions.InsufficientStockException;
 import com.uade.tpo.ecommerce.repository.CartItemRepository;
 import org.springframework.stereotype.Service;
 
@@ -13,11 +15,14 @@ public class CartItemService {
         this.repository = repository;
     }
 
-    public CartItem create(CartItem item) {
+    public CartItem create(CartItem item) throws InsufficientStockException {
+        int available = item.getProduct().getStock();
+        if (item.getQuantity() > available) throw new InsufficientStockException();
         return repository.save(item);
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws CartItemNotFoundException {
+        if (!repository.existsById(id)) throw new CartItemNotFoundException();
         repository.deleteById(id);
     }
 }
