+++ src/main/java/com/uade/tpo/ecommerce/service/ProductService.java
@@ -1,6 +1,8 @@
 package com.uade.tpo.ecommerce.service;
 
 import com.uade.tpo.ecommerce.entity.Product;
+import com.uade.tpo.ecommerce.exceptions.ProductDuplicateException;
+import com.uade.tpo.ecommerce.exceptions.ProductNotFoundException;
 import com.uade.tpo.ecommerce.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
@@ -19,20 +21,24 @@ public class ProductService {
         return repository.findAll();
     }
 
-    public Product getById(Long id) {
-        return repository.findById(id).orElse(null);
+    public Product getById(Long id) throws ProductNotFoundException {
+        return repository.findById(id).orElseThrow(ProductNotFoundException::new);
     }
 
-    public Product create(Product product) {
+    public Product create(Product product) throws ProductDuplicateException {
+        if (repository.existsByNameIgnoreCase(product.getName()))
+            throw new ProductDuplicateException();
         return repository.save(product);
     }
 
-    public Product update(Long id, Product product) {
+    public Product update(Long id, Product product) throws ProductNotFoundException {
+        if (!repository.existsById(id)) throw new ProductNotFoundException();
         product.setId(id);
         return repository.save(product);
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws ProductNotFoundException {
+        if (!repository.existsById(id)) throw new ProductNotFoundException();
         repository.deleteById(id);
     }
 }
